requirements.txt
asgiref==3.6.0
Django==4.2.7
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2023.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
black==23.3.0
drf-yasg==1.21.5
redis==4.5.5
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
